Auto routines to test PID and simulate cylce times
Added a PsuedoHatch routine to simulate the time spend getting or putting a hatch cover and a Wait() auto tool to be used as part of that.  The idea is drive between the loading station and rocket and then use PsuedoHatch() to simulate lining up then manipulating the hatch then backing up and turning around.  This is not tested but I don't want to risk losing it and I want to try other things.
Also merged in the TurnWithGyroPID routine that I tested but still needs tuning.