Added in OI improvements, duck controls, and PID position control for hatch lift.
Limit switches on hatch lift are not being used as of now, because no one seems to know how they're going to work, and it's possible they will be directly plugged into the motor controllers and I won't have to worry about them. I created a CommandGroup for extending the hatch, closing the finger, and retracting, so both controls for the duck, extending/retracting and raising finger/lowering finger, are all mapped to one button. Pressing the right bumper raises the finger, releasing runs the CommandGroup. The time delays between commands in the CommandGroup are not final,but it might be bad if everything decided to happen at once, so I put some in.