Uncommented code with hatch collector PID, and got rid of todo comments
I'm not sure why you'd want PID on an intake, but since there's code for getSelectedSensorPosition I might need to. The comments were no longer necessary, as I had gotten the work done (and Nathaniel wasn't the one who was supposed to work on it) so I got rid of them.