Fixed package names
I know that lowerCamelCase is the way we're *supposed* to do things, but it's not the way we *are* doing things, and if I change the folders I get some really bad conflicts, so I'm changing the package names back to what they were. In any case, the folders weren't renamed, so there was an error in every file I had to change, and now there isn't.