Added reverse mode
Because why not? Actually, it might be useful for testing controls for cameras, so I figured I'd add it in. Also made it so that if you're holding down one bumper (doesn't matter which) it goes at 75% speed, and if both are being held down it goes at 50%. Pressing X reverses controls.