Added selectable bumpers and left analog stick y-coordinate for operator control
Instead of using InstantCommand, there are methods in OI.java that are used in place of Robot.oi.operatorControl.[whatever].get(), and that only return true when the mode is what it should be (ex: Robot.oi.getLeftBumperCargo() returns the actual value only when Robot.operatorMode is "Cargo", and returns false otherwise. This avoids the movement of code from teleop commands to a separate command, and the only necessary changes are changing the access of the bumpers and left stick y-coordinate in each program, from the direct access to the new function.